Changes made to allow dosing via events
New additions are the Oral_dose, Drink_dose, Oral_on, Dermal_dose and Dermal_on compartments. The '_dose' compartments can obtain a specified dose via an event, whereas the '_on' compartments can be switched on and off via an event.
The exposure via milk during the first year is adjusted to periods of 3 months instead of per month, because otherwise the model crashes.